refactor: created auth package
