+++ backend/infrahub/auth/__init__.py
@@ -1,0 +1,1 @@
+from infrahub.auth.auth import *  # noqa: F403
