+++ backend/infrahub/auth/auth_groups/__init__.py
@@ -1,0 +1,7 @@
+from infrahub.auth.auth_groups.filter import ClaimFilter
+from infrahub.auth.auth_groups.service import AutoCreatedGroupsService
+
+__all__ = [
+    "AutoCreatedGroupsService",
+    "ClaimFilter",
+]
